Fix for NPEs in GeometryAdapter::buildGeometry. GeometryAdapter has moved from private nested class in order to make the lib more flexible in usage. Tests for new code branches.
