+++ core/src/main/java/com/github/filosganga/geogson/gson/GeometryAdapterFactory.java
@@ -16,19 +16,14 @@
 
 package com.github.filosganga.geogson.gson;
 
-import com.github.filosganga.geogson.model.*;
-import com.github.filosganga.geogson.model.positions.*;
+import com.github.filosganga.geogson.model.Feature;
+import com.github.filosganga.geogson.model.FeatureCollection;
+import com.github.filosganga.geogson.model.Geometry;
+import com.github.filosganga.geogson.model.positions.Positions;
 import com.google.gson.Gson;
 import com.google.gson.TypeAdapter;
 import com.google.gson.TypeAdapterFactory;
 import com.google.gson.reflect.TypeToken;
-import com.google.gson.stream.JsonReader;
-import com.google.gson.stream.JsonToken;
-import com.google.gson.stream.JsonWriter;
-
-import java.io.IOException;
-import java.util.LinkedList;
-import java.util.Optional;
 
 /**
  * The Gson TypeAdapterFactory responsible to serialize/de-serialize all the {@link Geometry}, {@link Feature}
@@ -50,148 +45,5 @@ public final class GeometryAdapterFactory implements TypeAdapterFactory {
         } else {
             return null;
         }
-    }
-
-    private static class GeometryAdapter extends TypeAdapter<Geometry> {
-
-        private final TypeAdapter<Geometry> geometryAdapter;
-        private final TypeAdapter<Positions> positionsAdapter;
-
-        private GeometryAdapter(Gson gson) {
-            this.geometryAdapter = gson.getAdapter(Geometry.class);
-            this.positionsAdapter = gson.getAdapter(Positions.class);
-        }
-
-        @Override
-        public void write(JsonWriter out, Geometry value) throws IOException {
-            if (value == null) {
-                out.nullValue();
-            } else {
-                out.beginObject();
-
-                out.name("type").value(value.type().getValue());
-                if (value.type() == Geometry.Type.GEOMETRY_COLLECTION) {
-                    out.name("geometries");
-                    out.beginArray();
-                    GeometryCollection geometries = (GeometryCollection) value;
-                    for (Geometry<?> geometry : geometries.getGeometries()) {
-                        geometryAdapter.write(out, geometry);
-                    }
-                    out.endArray();
-                } else {
-                    out.name("coordinates");
-                    positionsAdapter.write(out, value.positions());
-                }
-                out.endObject();
-            }
-        }
-
-        @Override
-        public Geometry<?> read(JsonReader in) throws IOException {
-
-            Geometry<?> geometry = null;
-            if (in.peek() == JsonToken.NULL) {
-                in.nextNull();
-            } else if (in.peek() == JsonToken.BEGIN_OBJECT) {
-                in.beginObject();
-
-                Geometry.Type type = null;
-                Positions positions = null;
-                Geometry<?> geometries = null;
-
-                while (in.hasNext()) {
-                    String name = in.nextName();
-                    if ("type".equals(name)) {
-                        type = Geometry.Type.forValue(in.nextString());
-                    } else if ("coordinates".equals(name)) {
-                        positions = positionsAdapter.read(in);
-                    } else if ("geometries".equals(name)) {
-                        geometries = readGeometries(in);
-                    } else {
-                        in.skipValue();
-                    }
-                }
-
-                geometry = buildGeometry(type, positions, geometries);
-
-                in.endObject();
-
-            } else {
-                throw new IllegalArgumentException("The given json is not a valid Geometry: " + in.peek());
-            }
-
-            return geometry;
-        }
-
-
-        private Geometry<?> readGeometries(JsonReader in) throws IOException {
-            Geometry<?> parsed;
-
-            JsonToken peek = in.peek();
-            if (peek == JsonToken.NULL) {
-                in.nextNull();
-                parsed = null;
-            } else if (peek == JsonToken.BEGIN_ARRAY) {
-                parsed = parseGeometries(in);
-            } else {
-                throw new IllegalArgumentException("The json must be an array or null: " + in.peek());
-            }
-
-            return parsed;
-        }
-
-        private Geometry<?> parseGeometries(JsonReader in) throws IOException {
-
-            Optional<Geometry<?>> parsed = Optional.empty();
-
-            if (in.peek() != JsonToken.BEGIN_ARRAY) {
-                throw new IllegalArgumentException("The given json is not a valid GeometryCollection");
-            }
-
-            in.beginArray();
-            if (in.peek() == JsonToken.BEGIN_OBJECT) {
-                LinkedList<Geometry<?>> geometries = new LinkedList<>();
-                while (in.hasNext()) {
-                    @SuppressWarnings("rawtypes")
-                    Geometry geometry = geometryAdapter.read(in);
-                    geometries.add(geometry);
-                }
-                parsed = Optional.<Geometry<?>>of(GeometryCollection.of(geometries));
-            }
-
-            in.endArray();
-
-            return parsed.orElse(null);
-        }
-
-        private Geometry<?> buildGeometry(final Geometry.Type type, final Positions positions, final Geometry<?> geometries) {
-            switch (type) {
-                case GEOMETRY_COLLECTION:
-                    return geometries;
-                case MULTI_POLYGON:
-                    return new MultiPolygon((MultiDimensionalPositions) positions);
-                case POLYGON:
-                    return new Polygon((AreaPositions) positions);
-                case MULTI_LINE_STRING:
-                    return new MultiLineString((AreaPositions) positions);
-                case LINEAR_RING:
-                case LINE_STRING:
-                    return ((LinearPositions) positions).isClosed()
-                            ? new LinearRing((LinearPositions) positions)
-                            : new LineString((LinearPositions) positions);
-                case MULTI_POINT:
-                    if (positions instanceof SinglePosition) {
-                        return new MultiPoint(LinearPositions.builder().addSinglePosition((SinglePosition) positions).build());
-                    } else {
-                        return new MultiPoint((LinearPositions) positions);
-                    }
-                case POINT:
-                    return new Point(((SinglePosition) positions));
-                default:
-                    throw new IllegalArgumentException("Cannot build a geometry for type: " + type);
-            }
-        }
-
-
     }
 }
